Moved LoginPage to an Authentication folder, made all routes protected
